Addon-docs: Rename Preview/Props to Canvas/ArgsTable
