+++ addons/docs/src/mdx/__testfixtures__/previews.output.snapshot
@@ -5,7 +5,7 @@ exports[`docs-mdx-compiler-plugin previews.mdx 1`] = `
 import { assertIsFn, AddContext } from '@storybook/addon-docs/blocks';
 
 import { Button } from '@storybook/react/demo';
-import { Preview, Story, Meta } from '@storybook/addon-docs/blocks';
+import { Canvas, Story, Meta } from '@storybook/addon-docs/blocks';
 
 const makeShortcode = (name) =>
   function MDXDefaultShortcode(props) {
@@ -30,9 +30,9 @@ function MDXContent({ components, ...props }) {
         }}
         mdxType=\\"Meta\\"
       />
-      <h1>{\`Preview\`}</h1>
-      <p>{\`Previews can contain normal components, stories, and story references\`}</p>
-      <Preview mdxType=\\"Preview\\">
+      <h1>{\`Canvas\`}</h1>
+      <p>{\`Canvases can contain normal components, stories, and story references\`}</p>
+      <Canvas mdxType=\\"Canvas\\">
         <Button mdxType=\\"Button\\">Just a button</Button>
         <Story name=\\"hello button\\" mdxType=\\"Story\\">
           <Button mdxType=\\"Button\\">Hello button</Button>
@@ -41,11 +41,11 @@ function MDXContent({ components, ...props }) {
           <Button mdxType=\\"Button\\">Two</Button>
         </Story>
         <Story id=\\"welcome--welcome\\" mdxType=\\"Story\\" />
-      </Preview>
-      <p>{\`Preview wthout a story\`}</p>
-      <Preview mdxSource=\\"%0A%3CButton%3EJust%20a%20button%3C/Button%3E%0A\\" mdxType=\\"Preview\\">
+      </Canvas>
+      <p>{\`Canvas wthout a story\`}</p>
+      <Canvas mdxSource=\\"%0A%3CButton%3EJust%20a%20button%3C/Button%3E%0A\\" mdxType=\\"Canvas\\">
         <Button mdxType=\\"Button\\">Just a button</Button>
-      </Preview>
+      </Canvas>
     </MDXLayout>
   );
 }
