+++ examples/aurelia-kitchen-sink/src/stories/mdx.stories.mdx
@@ -1,4 +1,4 @@
-import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
 import { CoolButton } from '../cool-button/cool-button';
 import { addComponents } from '@storybook/aurelia';
 import { text, withKnobs } from '@storybook/addon-knobs';
@@ -8,9 +8,9 @@ import Button from './button';
 
 # Welcome
 
-This is a test document written in MDX that defines a `CoolButton` story wrapped in a `Preview` doc block:
+This is a test document written in MDX that defines a `CoolButton` story wrapped in a `Canvas` doc block:
 
-<Preview withToolbar>
+<Canvas withToolbar>
   <Story name="to Storybook">
     {{
       template: `<cool-button text.bind="testText"></cool-button>`,
@@ -19,4 +19,4 @@ This is a test document written in MDX that defines a `CoolButton` story wrapped
       },
     }}
   </Story>
-</Preview>
+</Canvas>
