+++ examples/cra-ts-kitchen-sink/src/stories/Test.stories.mdx
@@ -1,18 +1,18 @@
-import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
 import { Button } from './Button';
 
 <Meta title="Test" />
 
 Here's some _markdown_!
 
-# Preview
+# Canvas
 
-<Preview>
+<Canvas>
   <Story name="button">
     <Button>hello</Button>
   </Story>
-</Preview>
+</Canvas>
 
-# Props
+# ArgsTable
 
-<Props of={Button} />
+<ArgsTable of={Button} />
