+++ examples/cra-ts-kitchen-sink/src/stories/docgen-tests/jsdoc/jsdoc-perfo.stories.mdx
@@ -1,10 +1,10 @@
-import { Meta, Props } from '@storybook/addon-docs/blocks';
+import { Meta, ArgsTable } from '@storybook/addon-docs/blocks';
 import { ButtonTooManyProps } from './jsdoc-perfo';
 
 <Meta title="Docgen/jsdoc-perfo" />
 
 ## Render 150 props with JSDoc tags
 
-<Props of={ButtonTooManyProps} />
-<Props of={ButtonTooManyProps} />
-<Props of={ButtonTooManyProps} />
+<ArgsTable of={ButtonTooManyProps} />
+<ArgsTable of={ButtonTooManyProps} />
+<ArgsTable of={ButtonTooManyProps} />
