+++ examples/cra-ts-kitchen-sink/src/stories/docgen-tests/types/types.stories.mdx
@@ -1,4 +1,4 @@
-import { Meta, Props } from '@storybook/addon-docs/blocks';
+import { Meta, ArgsTable } from '@storybook/addon-docs/blocks';
 import { PropTypesProps } from './prop-types';
 import { TypeScriptProps, TypeScriptHtmlComponent } from './ts-types';
 
@@ -6,10 +6,10 @@ import { TypeScriptProps, TypeScriptHtmlComponent } from './ts-types';
 
 ## Prop Types
 
-<Props of={PropTypesProps} />
+<ArgsTable of={PropTypesProps} />
 
 ## TypeScript
 
-<Props of={TypeScriptProps} />
+<ArgsTable of={TypeScriptProps} />
 
-<Props of={TypeScriptHtmlComponent} />
+<ArgsTable of={TypeScriptHtmlComponent} />
