+++ examples/cra-ts-kitchen-sink/src/stories/props-sort.stories.mdx
@@ -1,7 +1,7 @@
-import { PropsSort } from "./PropsSort";
-import { Props, Meta } from '@storybook/addon-docs/blocks';
+import { PropsSort } from './PropsSort';
+import { ArgsTable, Meta } from '@storybook/addon-docs/blocks';
 
 <Meta title="PropsSort" />
 
-<Props of={PropsSort} />
-<Props of={PropsSort} exclude={["foo"]} />
+<ArgsTable of={PropsSort} />
+<ArgsTable of={PropsSort} exclude={['foo']} />
