+++ examples/html-kitchen-sink/stories/addon-docs.stories.mdx
@@ -1,4 +1,4 @@
-import { Story, Preview, Meta } from '@storybook/addon-docs/blocks';
+import { Story, Canvas, Meta } from '@storybook/addon-docs/blocks';
 import { action } from '@storybook/addon-actions';
 
 # Storybook Docs for HTML
@@ -27,15 +27,15 @@ How you like them apples?!
 
 ## Custom source
 
-<Preview>
+<Canvas>
   <Story name="custom source" height="100px" parameters={{ docs: { source: { code: 'hello' } } }}>
     {'<h1>Custom source</h1>'}
   </Story>
-</Preview>
+</Canvas>
 
 ## Transformed source
 
-<Preview>
+<Canvas>
   <Story
     name="transformed source"
     height="100px"
@@ -43,7 +43,7 @@ How you like them apples?!
   >
     {'<h1>Some source</h1>'}
   </Story>
-</Preview>
+</Canvas>
 
 ## Story reference
