+++ examples/official-storybook/stories/addon-docs/addon-docs.stories.mdx
@@ -1,7 +1,7 @@
 import {
   Story,
-  Preview,
-  Props,
+  Canvas,
+  ArgsTable,
   Source,
   Description,
   ColorPalette,
@@ -33,11 +33,11 @@ import MdxNotes from '../notes/notes.mdx';
 
 <Source id="." />
 
-<Props of="." />
+<ArgsTable of="." />
 
-<Preview>
+<Canvas>
   <Story id="nonexistent-story" />
-</Preview>
+</Canvas>
 
 ## Transclusion
 
@@ -79,7 +79,7 @@ export const nonStory1 = 'foo'; // a non-story export
 
 export const nonStory2 = () => <Button>Not a story</Button>; // another one
 
-<Preview>
+<Canvas>
   <Story name="hello story">
     <Button onClick={action('clicked')}>{text('caption', 'hello world')}</Button>
   </Story>
@@ -93,7 +93,7 @@ export const nonStory2 = () => <Button>Not a story</Button>; // another one
     <Button onClick={action('clicked')}>goodbye world</Button>
   </Story>
   <Story name="plaintext">Plain text</Story>
-</Preview>
+</Canvas>
 
 <Story name="solo story">
   <Button onClick={action('clicked')}>solo</Button>
@@ -122,15 +122,15 @@ export const nonStory2 = () => <Button>Not a story</Button>; // another one
 
 <Source name="hello story" />
 
-<Preview>
-  <Button onClick={action('clicked')}>view source in preview</Button>
-</Preview>
+<Canvas>
+  <Button onClick={action('clicked')}>view source in Canvas</Button>
+</Canvas>
 
 ## Configurable height
 
-<Preview withToolbar>
+<Canvas withToolbar>
   <Story id="basics-button--all-buttons" height="400px" />
-</Preview>
+</Canvas>
 
 ## Description
 
@@ -138,29 +138,29 @@ export const nonStory2 = () => <Button>Not a story</Button>; // another one
 
 <Description of={DocgenButton} />
 
-## Props
+## ArgsTable
 
 ### Docgen
 
-<Props of={DocgenButton} />
+<ArgsTable of={DocgenButton} />
 
 ### Flow
 
 Flow types are not officially supported
 
-<Props of={FlowTypeButton} />
+<ArgsTable of={FlowTypeButton} />
 
 ### withSource="none"
 
-<Preview withSource="none">
+<Canvas withSource="none">
   <h1>no source</h1>
-</Preview>
+</Canvas>
 
-<Preview withSource="none">
-  <Story name="no source preview">
+<Canvas withSource="none">
+  <Story name="no source Canvas">
     <Button onClick={action('clicked')}>solo</Button>
   </Story>
-</Preview>
+</Canvas>
 
 ## Docs Disable
 
@@ -170,9 +170,9 @@ Flow types are not officially supported
 
 ## Story with multiple children
 
-<Preview>
+<Canvas>
   <Story name="story multiple children">
     <p>Hello Child #1</p>
     <p>Hello Child #2</p>
   </Story>
-</Preview>
+</Canvas>
