+++ examples/official-storybook/stories/demo/button.stories.mdx
@@ -1,13 +1,13 @@
 import { useState } from 'react';
 import { action } from '@storybook/addon-actions';
 import { Button } from '@storybook/react/demo';
-import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
 
 <Meta title="Other/Demo/ButtonMdx" component={Button} />
 
 # Simple Button
 
-<Props of={Button} />
+<ArgsTable of={Button} />
 
 ## Hello
 
@@ -27,7 +27,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 
 ## Counter w/ Code
 
-<Preview>
+<Canvas>
   <Story name="with counter">
     {() => {
       const [counter, setCounter] = useState(0);
@@ -35,4 +35,4 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
       return <Button onClick={() => setCounter(counter + 1)}>{label}</Button>;
     }}
   </Story>
-</Preview>
+</Canvas>
