+++ examples/official-storybook/stories/notes/notes.mdx
@@ -1,4 +1,4 @@
-import { Props, Story } from '@storybook/addon-docs/blocks';
+import { ArgsTable, Story } from '@storybook/addon-docs/blocks';
 import { Button } from '@storybook/react/demo';
 
 # Welcome!
@@ -7,6 +7,6 @@ Let's write markdown without stories
 
 But we can still use doc blocks in our stories
 
-<Props of={Button} />
+<ArgsTable of={Button} />
 
 <Story id="basics-button--all-buttons" height="400px" />
