+++ examples/vue-kitchen-sink/src/stories/addon-docs.stories.mdx
@@ -1,4 +1,4 @@
-import { Story, Preview, Meta } from '@storybook/addon-docs/blocks';
+import { Story, Canvas, Meta } from '@storybook/addon-docs/blocks';
 import { stripIndents } from 'common-tags';
 import MyButton from './Button.vue';
 
@@ -68,13 +68,13 @@ The default configuration for `addon-docs/vue` implements the `prepareForInline`
 
 To provide a tangible example of the inline/iframe rendering contexts, here's the same story rendered both ways.
 
-<Preview>
+<Canvas>
   <Story id="app--app" inline={true} />
-</Preview>
+</Canvas>
 
-<Preview>
+<Canvas>
   <Story id="app--app" height="200px" />
-</Preview>
+</Canvas>
 
 The biggest win here is that we don't have to worry about setting the height anymore. The story can calculate its height like any other inline element. As you can see, several of the stories on this page are quite small, but take up `500px`, because that's the default story height, and we didn't tweak the story to be an explicit height. Another huge gain here is in terms of performance. Using an iframe to render a single element is definitely unnecessary in most cases and, as you may have noticed while reading this page, can _really_ cause page performance to suffer.
 
@@ -88,7 +88,7 @@ Just like in React, we can easily reference other stories in our docs:
 
 ## More info
 
-For more info, check out the [Storybook Docs Technical Preview](https://docs.google.com/document/d/1un6YX7xDKEKl5-MVb-egnOYN8dynb5Hf7mq0hipk8JE/edit?usp=sharing).
+For more info, check out the [Storybook Docs Technical Canvas](https://docs.google.com/document/d/1un6YX7xDKEKl5-MVb-egnOYN8dynb5Hf7mq0hipk8JE/edit?usp=sharing).
 
 We want your feedback to help make this more useful.
